BMP Sensor implementation and cleaning up of logging
I added BMP sensor code for (hopefully actually working) pressure/temp data-gathering.
Logging now just writes to a struct, which can then be passed to a secondary function which generates a string.
This is cleaner now debugging is less necessary and the code is less experimental.
N.B. Claude Sonnet was used to write most of this code so not my fault if it's broken in some way